feat: implement order and checkout

+++ src/main/java/com/techstore/laptop_shop/domain/CartDetail.java
@@ -16,7 +16,7 @@ public class CartDetail {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private long id;
 
-    private long quantity;
+    private int quantity;
 
     private double price;
 
@@ -38,11 +38,11 @@ public class CartDetail {
         this.id = id;
     }
 
-    public long getQuantity() {
+    public int getQuantity() {
         return quantity;
     }
 
-    public void setQuantity(long quantity) {
+    public void setQuantity(int quantity) {
         this.quantity = quantity;
     }
